project/The project back/requirements.txt
Django==6.0
djangorestframework==3.14.0
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.2
Pillow==10.1.0
python-decouple==3.8
